chore: spliting builder into different classes with specific context
